Get rid of runtime/capi.h
