+++ src/capi/errors.cpp
@@ -21,7 +21,6 @@
 #include "core/types.h"
 #include "core/util.h"
 #include "gc/collector.h"
-#include "runtime/capi.h"
 #include "runtime/dict.h"
 #include "runtime/objmodel.h"
 #include "runtime/types.h"
