+++ src/codegen/irgen/hooks.cpp
@@ -38,7 +38,6 @@
 #include "core/stats.h"
 #include "core/types.h"
 #include "core/util.h"
-#include "runtime/capi.h"
 #include "runtime/objmodel.h"
 #include "runtime/types.h"
