+++ src/codegen/pypa-parser.cpp
@@ -33,7 +33,6 @@
 #include "core/types.h"
 #include "core/util.h"
 #include "gc/collector.h"
-#include "runtime/capi.h"
 #include "runtime/objmodel.h"
 #include "runtime/types.h"
