--- src/runtime/capi.h
@@ -1,40 +1,0 @@
-// Copyright (c) 2014-2015 Dropbox, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef PYSTON_RUNTIME_CAPI_H
-#define PYSTON_RUNTIME_CAPI_H
-
-#include <string>
-
-namespace pyston {
-
-class Box;
-class BoxedModule;
-BoxedModule* importCExtension(const std::string& full_name, const std::string& last_name, const std::string& path);
-
-void checkAndThrowCAPIException();
-void throwCAPIException() __attribute__((noreturn));
-struct ExcInfo;
-void setCAPIException(const ExcInfo& e);
-
-#define fatalOrError(exception, message)                                                                               \
-    do {                                                                                                               \
-        if (CONTINUE_AFTER_FATAL)                                                                                      \
-            PyErr_SetString((exception), (message));                                                                   \
-        else                                                                                                           \
-            Py_FatalError((message));                                                                                  \
-    } while (0)
-}
-
-#endif
