+++ src/runtime/import.h
@@ -21,6 +21,7 @@ namespace pyston {
 
 extern "C" Box* import(int level, Box* from_imports, llvm::StringRef module_name);
 extern Box* importModuleLevel(llvm::StringRef module_name, Box* globals, Box* from_imports, int level);
+BoxedModule* importCExtension(const std::string& full_name, const std::string& last_name, const std::string& path);
 }
 
 #endif
