+++ src/runtime/inline/xrange.cpp
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 #include "core/types.h"
-#include "runtime/capi.h"
 #include "runtime/objmodel.h"
 #include "runtime/types.h"
