+++ src/runtime/int.cpp
@@ -24,7 +24,6 @@
 #include "core/stats.h"
 #include "core/types.h"
 #include "gc/collector.h"
-#include "runtime/capi.h"
 #include "runtime/float.h"
 #include "runtime/inline/boxing.h"
 #include "runtime/long.h"
